refactor: move tests out of src into top-level tests/
Mirror the src/ package layout under tests/ (dropping the per-module specs/
segment) so the package code stays clean and tests don't sit inside src/.
Imports rewritten to reference ../src/...; vitest include -> tests/**/*.test.ts.
