+++ tests/review/instructions.test.ts
@@ -2,8 +2,8 @@ import { mkdtemp, rm, writeFile } from 'node:fs/promises'
 import { tmpdir } from 'node:os'
 import { join } from 'node:path'
 import { afterEach, beforeEach, describe, expect, test } from 'vitest'
-import type { ReviewConfig } from '../config'
-import { buildInstructions } from '../instructions'
+import type { ReviewConfig } from '../../src/review/config'
+import { buildInstructions } from '../../src/review/instructions'
 
 const makeConfig = (
   workspace: string,
